Transplant of Two grid stuff out of cross-sections

+++ ChiModules/LinearBoltzmannSolver/SweepChunks/lbs_sweepchunk_pwl2.cc
@@ -66,7 +66,7 @@ void lbs::SweepChunkPWL::
   const SubSetInfo& grp_ss_info = groupset.grp_subset_infos[angle_set->ref_subset];
   const size_t gs_ss_size  = grp_ss_info.ss_size;
   const size_t gs_ss_begin = grp_ss_info.ss_begin;
-  const size_t gs_gi = groupset.groups[gs_ss_begin].id; // Groupset subset first group number
+  const int    gs_gi = groupset.groups[gs_ss_begin].id; // Groupset subset first group number
 
   int deploc_face_counter = -1;
   int preloc_face_counter = -1;
